renamed quote table components
