+++ src/pages/projEstimate/FormProjEstimate.tsx
@@ -8,7 +8,7 @@ import { ScrollToFieldError } from '../../components/utils/ScrollToFieldError';
 import { getFieldName, statusChoices, TypeOfForm } from './form';
 import SummaryTable from './SummaryTable/SummaryTable';
 
-import { RenderFunc } from './QuoteTable/RenderFunc';
+import { renderQuoteTable } from './QuoteTable/';
 import { SubTotalTable } from './SubTotalTable/SubTotalTable';
 import { useUpdateProjectId } from './hooks/useUpdateProjectId';
 import { SearchProject } from './fieldComponents/SearchProject';
@@ -62,7 +62,7 @@ export default function FormProjEstimate() {
           {/* 見積もり用のテーブル */}
           <FieldArray
             name={getFieldName('items')}
-            render={RenderFunc}
+            render={renderQuoteTable}
           />
         </Grid>
